upstream lb: move thread aware lb base to extensions/
