+++ source/extensions/load_balancing_policies/maglev/BUILD
@@ -17,7 +17,7 @@ envoy_cc_library(
         "//envoy/upstream:load_balancer_interface",
         "//source/common/common:bit_array_lib",
         "//source/common/runtime:runtime_features_lib",
-        "//source/common/upstream:thread_aware_lb_lib",
+        "//source/extensions/load_balancing_policies/common:thread_aware_lb_lib",
         "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/load_balancing_policies/maglev/v3:pkg_cc_proto",
     ],
@@ -35,7 +35,7 @@ envoy_cc_library(
         "//envoy/upstream:load_balancer_interface",
         "//source/common/common:bit_array_lib",
         "//source/common/runtime:runtime_features_lib",
-        "//source/common/upstream:thread_aware_lb_lib",
+        "//source/extensions/load_balancing_policies/common:thread_aware_lb_lib",
         "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/load_balancing_policies/maglev/v3:pkg_cc_proto",
     ],
