+++ source/extensions/load_balancing_policies/ring_hash/BUILD
@@ -19,7 +19,7 @@ envoy_cc_library(
     deps = [
         "//envoy/upstream:load_balancer_interface",
         "//source/common/common:minimal_logger_lib",
-        "//source/common/upstream:thread_aware_lb_lib",
+        "//source/extensions/load_balancing_policies/common:thread_aware_lb_lib",
         "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/load_balancing_policies/ring_hash/v3:pkg_cc_proto",
     ],
